Re Built the webapp using typescript enabled and also enabled tailwind css
