--- app/src/components/AnimatedLetters/AnimatedLetters.js
@@ -1,18 +1,0 @@
-import './AnimatedLetters.scss';
-
-const AnimatedLetters = ({ letterClass, stringArray, index }) => {
-  return (
-    <span>
-      {stringArray.map((char, i) => (
-        <span
-          key={char + i}
-          className={letterClass + ' _' + (i + index).toString()}
-        >
-          {char}
-        </span>
-      ))}
-    </span>
-  );
-};
-
-export default AnimatedLetters;
